Added bonuses
Creation of 4 new prefabs:
-3 different bonuses (speed upgrade, slow upgrade and ball split)
-1 new brick that summons a random bonus when broken.

Added a script "Display Bonus" to instantiate a random bonus.

Their effect are coded in the "Ball Manager Script"

Ratio Speed can be modified in Unity directly as well as the bonus brick ratio appearance.

this ratio has been coded in the "Brick Manager" script

May be not a bug :
It seems that when the ball splits, it can generate another ball, instead of generating only one.

If two balls are on a bonus at the same frame, it can double the effects of the bonus. 
Make sure to not set the ball's speed to high.